Add Auth, jwt, user password storage and hashing

+++ src/Controllers/UserController.cs
@@ -1,6 +1,6 @@
+using System.Security.Claims;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
-using System.Collections.Generic;
 
 namespace User.Controllers
 {
@@ -8,13 +8,11 @@ namespace User.Controllers
     [Route("api/v1/[controller]")]
     public class UserController : ControllerBase
     {
-        private readonly IUserService _userService;
-        // private readonly IPasswordHasher<object> _passwordHasher;
+        private readonly IUserProfileService _userService;
 
-        public UserController(IUserService userService)
+        public UserController(IUserProfileService userService)
         {
             _userService = userService;
-            // _passwordHasher = passwordHasher;
         }
 
         [HttpGet("{id}")]
@@ -44,7 +42,6 @@ namespace User.Controllers
         {
             try
             {
-
                 var result = await _userService.AddUserAsync(request);
                 return Created("User", result);
             }
@@ -59,9 +56,20 @@ namespace User.Controllers
         }
 
         [HttpPut("{id}")]
+        [Authorize]
         public async Task<ActionResult> UpdateUser(string id, [FromBody] UpdateUserRequest request)
         {
             request.Id = id;
+
+            var userId = User.HasClaim(c => c.Type == ClaimTypes.NameIdentifier) ? User.FindFirstValue(ClaimTypes.NameIdentifier) : null;
+            if (string.IsNullOrEmpty(userId))
+            {
+                return Unauthorized("User ID claim is missing.");
+            }
+            if (userId != id)
+            {
+                return Unauthorized("User can only update their own account!");
+            }
             try
             {
                 await _userService.UpdateUserAsync(request);
@@ -82,10 +90,20 @@ namespace User.Controllers
         }
 
         [HttpDelete("{id}")]
+        [Authorize]
         public async Task<ActionResult> DeleteUser(string id)
         {
             try
             {
+                var userId = User.HasClaim(c => c.Type == ClaimTypes.NameIdentifier) ? User.FindFirstValue(ClaimTypes.NameIdentifier) : null;
+                if (string.IsNullOrEmpty(userId))
+                {
+                    return Unauthorized("User ID claim is missing.");
+                }
+                if (userId != id)
+                {
+                    return Unauthorized("User can only delete their own account!");
+                }
                 await _userService.DeleteUserAsync(id);
                 return NoContent();
             }
@@ -102,52 +120,5 @@ namespace User.Controllers
                 return StatusCode(500, $"Internal server error: {ex.Message}");
             }
         }
-
-        // // Login
-        // [HttpPost("login")]
-        // public async Task<ActionResult<LoginResponse>> Login([FromBody] LoginRequest request)
-        // {
-        //     try
-        //     {
-        //         var response = await _userService.LoginAsync(request);
-        //         return Ok(response);
-
-        //     }
-        //     catch (ArgumentException ex)
-        //     {
-        //         return BadRequest(ex.Message);
-        //     }
-        //     catch (KeyNotFoundException ex)
-        //     {
-        //         return NotFound(ex.Message);
-        //     }
-        //     catch (Exception ex)
-        //     {
-        //         return StatusCode(500, $"Internal server error: {ex.Message}");
-        //     }
-        // }
-
-        // // logout
-        // [HttpPost("logout")]
-        // public async Task<ActionResult> Logout([FromBody] LogoutRequest request)
-        // {
-        //     try
-        //     {
-        //         await _userService.LogoutAsync(request);
-        //         return NoContent();
-        //     }
-        //     catch (ArgumentException ex)
-        //     {
-        //         return BadRequest(ex.Message);
-        //     }
-        //     catch (KeyNotFoundException ex)
-        //     {
-        //         return NotFound(ex.Message);
-        //     }
-        //     catch (Exception ex)
-        //     {
-        //         return StatusCode(500, $"Internal server error: {ex.Message}");
-        //     }
-        // }
     }
 }
